+++ hmdv/hmdv.cpp
@@ -109,7 +109,7 @@ static pmode mode2pmode(const mode selected)
         case mode::all:
             return pmode::all;
         default:
-            HMDQ_EXCEPTION(fmt::format("mode2pmode({}) is undefined", selected));
+            HMDQ_EXCEPTION(fmt::format("mode2pmode({}) is undefined", static_cast<int>(selected)));
     }
 }
